refactor: rename domain package to model
For its meaning and purpose, "model" sounds like more reasonable for this package
