+++ src/main/java/racingcar/model/CarName.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.model;
 
 public class CarName {
