+++ src/main/java/racingcar/model/Position.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.model;
 
 import java.util.Objects;
