+++ src/main/java/racingcar/view/OutputView.java
@@ -1,6 +1,6 @@
 package racingcar.view;
 
-import racingcar.domain.Car;
+import racingcar.model.Car;
 
 import java.util.List;
 import java.util.stream.Collectors;
