+++ src/test/java/model/CarMovementTest.java
@@ -1,10 +1,10 @@
-package domain;
+package model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.CarMovement;
+import racingcar.model.CarMovement;
 
 import static org.assertj.core.api.Assertions.*;
