+++ src/test/java/model/CarTest.java
@@ -1,11 +1,11 @@
-package domain;
+package model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.domain.Car;
-import racingcar.domain.CarMovement;
+import racingcar.model.Car;
+import racingcar.model.CarMovement;
 
 import java.util.List;
