+++ src/test/java/model/TrialCountTest.java
@@ -1,9 +1,9 @@
-package domain;
+package model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.TrialCount;
+import racingcar.model.TrialCount;
 
 import static org.assertj.core.api.Assertions.*;
